+++ spring-cloud-stream-samples/double/src/main/java/config/sink/SinkModuleDefinition.java
@@ -19,7 +19,7 @@ package config.sink;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.cloud.stream.annotation.EnableBinding;
-import org.springframework.cloud.stream.annotation.Sink;
+import org.springframework.cloud.stream.messaging.Sink;
 import org.springframework.integration.annotation.ServiceActivator;
 
 /**
