+++ spring-cloud-stream-samples/sink/src/main/java/demo/LogSink.java
@@ -19,7 +19,7 @@ package demo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.cloud.stream.annotation.EnableBinding;
-import org.springframework.cloud.stream.annotation.Sink;
+import org.springframework.cloud.stream.messaging.Sink;
 import org.springframework.integration.annotation.ServiceActivator;
 
 /**
