+++ spring-cloud-stream-samples/transform/src/main/java/demo/LoggingTransformer.java
@@ -20,7 +20,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.cloud.stream.annotation.EnableBinding;
-import org.springframework.cloud.stream.annotation.Processor;
+import org.springframework.cloud.stream.messaging.Processor;
 import org.springframework.integration.annotation.ServiceActivator;
 
 /**
