+++ spring-cloud-stream-test-support/src/main/java/org/springframework/cloud/stream/test/matcher/MessageQueueMatcher.java
@@ -38,7 +38,7 @@ import org.springframework.messaging.Message;
  * public class TransformProcessorApplicationTests {
  *
  *    {@literal @}Autowired
- *    {@literal @}ModuleChannels(TransformProcessor.class)
+ *    {@literal @}Bindings(TransformProcessor.class)
  *    private Processor processor;
  *
  *    {@literal @}Autowired
