+++ spring-cloud-stream-test-support/src/test/java/org/springframework/cloud/stream/test/ExampleTest.java
@@ -27,8 +27,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.test.IntegrationTest;
 import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.cloud.stream.annotation.EnableBinding;
-import org.springframework.cloud.stream.annotation.ModuleChannels;
-import org.springframework.cloud.stream.annotation.Processor;
+import org.springframework.cloud.stream.annotation.Bindings;
+import org.springframework.cloud.stream.messaging.Processor;
 import org.springframework.cloud.stream.test.binder.MessageCollector;
 import org.springframework.integration.annotation.Transformer;
 import org.springframework.messaging.Message;
@@ -47,7 +47,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 public class ExampleTest {
 
 	@Autowired
-	@ModuleChannels(MyProcessor.class)
+	@Bindings(MyProcessor.class)
 	private Processor processor;
 
 	@Autowired
