+++ spring-cloud-stream/src/main/java/org/springframework/cloud/stream/messaging/Sink.java
@@ -14,15 +14,24 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.stream.annotation;
+package org.springframework.cloud.stream.messaging;
 
-import org.springframework.messaging.MessageChannel;
+import org.springframework.cloud.stream.annotation.Input;
+import org.springframework.messaging.SubscribableChannel;
 
-public interface Source {
-	
-	public static String OUTPUT = "output";
-	
-	@Output(Source.OUTPUT)
-	MessageChannel output();
+/**
+ * Bindable interface with one input channel.
+ *
+ *
+ * @see org.springframework.cloud.stream.annotation.EnableBinding
+ * @author Dave Syer
+ * @author Marius Bogoevici
+ */
+public interface Sink {
+
+	String INPUT = "input";
+
+	@Input(Sink.INPUT)
+	SubscribableChannel input();
 
 }
