+++ spring-cloud-stream/src/test/java/org/springframework/cloud/stream/binder/ArbitraryInterfaceBindingTestsWithDefaults.java
@@ -30,7 +30,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.cloud.stream.annotation.EnableBinding;
-import org.springframework.cloud.stream.annotation.ModuleChannels;
+import org.springframework.cloud.stream.annotation.Bindings;
 import org.springframework.cloud.stream.utils.MockBinderConfiguration;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
@@ -43,7 +43,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 public class ArbitraryInterfaceBindingTestsWithDefaults {
 
 	@Autowired
-	@ModuleChannels(ArbitraryInterfaceBindingTestsWithDefaults.TestFooChannels.class)
+	@Bindings(ArbitraryInterfaceBindingTestsWithDefaults.TestFooChannels.class)
 	public FooChannels fooChannels;
 
 	@SuppressWarnings("rawtypes")
